Changed directory structure #24 * Moved tests and scraper code in app/
